Add PROXY protocol authority TLV helpers
